Extract everything related with the comment to its own component
